+++ ethergo/backends/mocks/simulated_test_backend.go
@@ -1253,6 +1253,27 @@ func (_m *SimulatedTestBackend) WaitForConfirmation(ctx context.Context, transac
 	_m.Called(ctx, transaction)
 }
 
+// Web3Version provides a mock function with given fields: ctx
+func (_m *SimulatedTestBackend) Web3Version(ctx context.Context) (string, error) {
+	ret := _m.Called(ctx)
+
+	var r0 string
+	if rf, ok := ret.Get(0).(func(context.Context) string); ok {
+		r0 = rf(ctx)
+	} else {
+		r0 = ret.Get(0).(string)
+	}
+
+	var r1 error
+	if rf, ok := ret.Get(1).(func(context.Context) error); ok {
+		r1 = rf(ctx)
+	} else {
+		r1 = ret.Error(1)
+	}
+
+	return r0, r1
+}
+
 type mockConstructorTestingTNewSimulatedTestBackend interface {
 	mock.TestingT
 	Cleanup(func())