+++ packages/contracts-communication/test/integration/PingPong.Dst.t.sol
@@ -88,7 +88,7 @@ contract PingPongDstIntegrationTest is PingPongIntegrationTest {
     function test_verifyRemoteBatch_state_db() public {
         module.verifyRemoteBatch(moduleBatch, moduleSignatures);
         skip(LONG_PERIOD);
-        assertEq(icDB.checkVerification(address(module), srcEntry, new bytes32[](0)), INITIAL_TS);
+        assertEq(icDB.checkBatchVerification(address(module), srcBatch), INITIAL_TS);
     }
 
     function test_interchainExecute_callPingPongApp() public {