+++ packages/contracts-communication/test/legacy/MessageBus.Management.t.sol
@@ -9,6 +9,7 @@ import {MessageBusBaseTest, InterchainClientV1Mock} from "./MessageBus.Base.t.so
 // solhint-disable ordering
 contract MessageBusManagementTest is MessageBusBaseTest {
     uint256 public constant LENGTH_ESTIMATE = 100;
+    uint64 public constant BUFFER = 12_345;
 
     function mockInterchainFees(uint256 length) internal {
         uint256 legacyMessageLen = LegacyMessageLib.payloadSize(length);
@@ -17,6 +18,29 @@ contract MessageBusManagementTest is MessageBusBaseTest {
             (REMOTE_CHAIN_ID, execService, icModules, icOptions, legacyMessageLen)
         );
         vm.mockCall(icClient, expectedCalldata, abi.encode(MOCK_FEE));
+    }
+
+    function test_gasBuffer_defaultValue() public {
+        assertEq(messageBus.gasBuffer(), GAS_BUFFER);
+    }
+
+    function test_setGasBuffer_emitsEvent() public {
+        expectEventGasBufferSet(BUFFER);
+        vm.prank(governor);
+        messageBus.setGasBuffer(BUFFER);
+    }
+
+    function test_setGasBuffer_setsBuffer() public {
+        vm.prank(governor);
+        messageBus.setGasBuffer(BUFFER);
+        assertEq(messageBus.gasBuffer(), BUFFER);
+    }
+
+    function test_setGasBuffer_revert_notGovernor(address caller) public {
+        vm.assume(caller != governor);
+        expectRevertUnauthorizedGovernor(caller);
+        vm.prank(caller);
+        messageBus.setGasBuffer(BUFFER);
     }
 
     function test_setMessageLengthEstimate_emitsEvent() public {